requirements.txt
httplib2==0.9.2
Flask==0.10.1
Flask-Admin==1.1.0
Flask-RESTful==0.3.5
Flask-Login==0.3.2
Flask-OAuthLib==0.9.2
flask-mail==0.9.1
flask-sslify==0.1.5
passlib==1.6.2
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.3
aniso8601==1.1.0
blinker==1.3
itsdangerous==0.24
peewee==2.6.1
psycopg2==2.5.2
python-dateutil==2.4.2
pytz==2015.7
redis==2.10.5
requests==2.3.0
six==1.10.0
sqlparse==0.1.8
wsgiref==0.1.2
Flask-Script==0.6.6
honcho==0.5.0
statsd==2.1.2
gunicorn==19.4.5
celery==3.1.11
jsonschema==2.4.0
click==3.3
RestrictedPython==3.6.0
wtf-peewee==0.2.3
pysaml2==2.4.0
pycrypto==2.6.1
funcy==1.5
raven==5.9.2
semver==2.2.1
xlsxwriter==0.8.4
pystache==0.5.4
parsedatetime==2.1
cryptography==1.4
inflector==2.0.11
